move: merge를 위한 사전 작업
