+++ src/main/java/chzzk/grassdiary/domain/base/BaseTimeEntity.java
@@ -2,16 +2,15 @@ package chzzk.grassdiary.domain.base;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
 import lombok.Getter;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import java.time.LocalDateTime;
-
 @Getter
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public class BaseTimeEntity extends BaseCreatedTimeEntity{
+public class BaseTimeEntity extends BaseCreatedTimeEntity {
     @LastModifiedDate
     private LocalDateTime updatedAt;
 }
