+++ src/main/java/chzzk/grassdiary/domain/diary/DiaryLike.java
@@ -1,7 +1,7 @@
 package chzzk.grassdiary.domain.diary;
 
 import chzzk.grassdiary.domain.base.BaseCreatedTimeEntity;
-import chzzk.grassdiary.domain.member.entity.Member;
+import chzzk.grassdiary.domain.member.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
