+++ src/main/java/chzzk/grassdiary/service/MyPageService.java
@@ -4,7 +4,7 @@ import chzzk.grassdiary.domain.color.ColorCode;
 import chzzk.grassdiary.domain.diary.Diary;
 import chzzk.grassdiary.domain.diary.DiaryRepository;
 import chzzk.grassdiary.domain.member.Member;
-import chzzk.grassdiary.domain.member.MemberRepository;
+import chzzk.grassdiary.domain.member.repository.MemberRepository;
 import chzzk.grassdiary.web.dto.member.GrassInfoDTO;
 import chzzk.grassdiary.web.dto.member.MemberInfoDTO;
 import chzzk.grassdiary.web.dto.member.TotalRewardDTO;
@@ -23,7 +23,7 @@ public class MyPageService {
     public MemberInfoDTO findProfileById(Long id) {
         Member member = memberRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 멤버 입니다. (id: " + id + ")"));
-        return new MemberInfoDTO(member.getProfileImageUrl(), member.getNickname(), member.getProfileIntro());
+        return new MemberInfoDTO(member.getPicture(), member.getNickname(), member.getProfileIntro());
     }
 
     @Transactional(readOnly = true)
