+++ src/test/java/chzzk/grassdiary/service/diary/DiaryServiceTest.java
@@ -5,7 +5,7 @@ import chzzk.grassdiary.domain.color.ColorCodeRepository;
 import chzzk.grassdiary.domain.color.ConditionLevel;
 import chzzk.grassdiary.domain.diary.DiaryRepository;
 import chzzk.grassdiary.domain.member.Member;
-import chzzk.grassdiary.domain.member.MemberRepository;
+import chzzk.grassdiary.domain.member.repository.MemberRepository;
 import chzzk.grassdiary.web.dto.diary.DiarySaveDTO;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
