+++ src/test/java/chzzk/grassdiary/service/diary/TagServiceTest.java
@@ -10,7 +10,7 @@ import chzzk.grassdiary.domain.diary.tag.MemberTagsRepository;
 import chzzk.grassdiary.domain.diary.tag.TagList;
 import chzzk.grassdiary.domain.diary.tag.TagListRepository;
 import chzzk.grassdiary.domain.member.Member;
-import chzzk.grassdiary.domain.member.MemberRepository;
+import chzzk.grassdiary.domain.member.repository.MemberRepository;
 import chzzk.grassdiary.web.dto.diary.TagDTO;
 import java.util.ArrayList;
 import java.util.List;
@@ -23,23 +23,20 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class TagServiceTest {
 
-    @Autowired
-    MemberRepository memberRepository;
-    @Autowired
-    DiaryRepository diaryRepository;
-
-    @Autowired
-    MemberTagsRepository memberTagsRepository;
-    @Autowired
-    TagListRepository tagListRepository;
-
-    @Autowired
-    TagService tagService;
-
     protected List<Member> members;
     protected List<Diary> diaries;
     protected List<MemberTags> memberTags;
     protected List<TagList> tagLists;
+    @Autowired
+    MemberRepository memberRepository;
+    @Autowired
+    DiaryRepository diaryRepository;
+    @Autowired
+    MemberTagsRepository memberTagsRepository;
+    @Autowired
+    TagListRepository tagListRepository;
+    @Autowired
+    TagService tagService;
 
     @BeforeEach
     void beforeEach() {
